ex_slp v0.1.2 ExSlp.Server

Functions

deregister(service)

Deregisters the service which has been registered earlier. You should call this method every time the application is about to be terminated. The service won’t be automatically deregistered. Takes service as a mandatory argument, the same you used to register the service. args is a standard openslp keyword list. Returns:

  { :ok, resp } # in case of success
  { :error, reason } # otherwise

deregister(service, args)

register(service)

Registers the service in the local network. Takes the service specification as a mandatory argument, args and opts as a standard keyword lists (see ExSlp.Client.findsrvs for more info on args. This is the place you should specify the service attributes. Check the original documentation for more service internals: http://www.openslp.org/doc/html/ProgrammersGuide/SLPReg.html Please keep in mind the slpd instance won’t track the initiator status and moreover it knows nothing about it. It’s initiator’s responsibility to deregister the service it registers. Returns:

{ :ok, resp } # in case of success,
{ :error, message } # otherwise.

Example:

register( "myservice://192.168.0.10" )
register( "service:myservice.xyz://192.168.0.10", [ attr1: val1, att2: val2 ] )
register( "service:myservice.xyz://192.168.0.10", [ l: "en", t: 60 * 60], [] )

register(service, opts)

register(service, args, opts)

status()

Checks the status of slpd daemon. Returns:

{ :ok, pid } # in case of success,
{ :not_running, message } # otherwise.