ex_slp v0.1.2 ExSlp.Util

Functions

compact(list)

Compacts an Enumerable instance.

format_args(args)

Takes a keyword list and formats it as a command line arguments like: [ "-arg1", "val1", "-arg2", "val2", ... ] Known args: slptool v1: v(ersion), s(cope), l(anguage) slptool v2: v(ersion), s(cope), l(anguage), t(ime), i(nterfaces), u(nicastic) Unknown args would be dropped silently.

format_opts(opts)

Takes a keyword list and formats it as slptool opts list: “(arg1=val1),(arg2=val2)”

format_servise_url(service)

Formats the service url and makes it look like: service:myservice://…

parse_opts(opts_s)

Takes a string of slptool opts and returns a keyword list.

parse_url(url)

The normal URI.parse which truncates the “service:” preffix and the lifetime postfix.